Spring boot service provides library using google mail service for free.
Email notification system which is using JavaMailSender from org.springframework.boot (spring-boot-starter-maven's dependency) i've created 1 simple class NotificationService which will be later used in GenericRestController for the message which we are going to send.

Once we create the other web app services we will be using mail notifications to notify our users for events or declined appointments.That will be easily imported.